Make Simulator a template class
