+++ test/simulator/utils.hpp
@@ -12,7 +12,7 @@ namespace test {
 using two_way_link_t = std::pair<std::shared_ptr<sim::IRoutingDevice>,
                                  std::shared_ptr<sim::IRoutingDevice>>;
 
-void add_two_way_links(sim::Simulator& sim,
+void add_two_way_links(sim::BasicSimulator& sim,
                std::initializer_list<two_way_link_t> links);
 
 bool check_reachability(std::shared_ptr<sim::IRoutingDevice> src_device,
